+++ components/map/MapTooltip.js
@@ -3,6 +3,8 @@ import { useSelector } from "react-redux";
 
 export default function MapTooltip() {
   const currentHoverTarget = useSelector((state) => state.currentHoverTarget);
+  const widgetLocations = useSelector(state => state.widgetLocations);
+  const widgetsOnRight = widgetLocations.some(o => o.side === "right");
   
   if (!(typeof window) || !currentHoverTarget.data) return null;
 
@@ -16,9 +18,9 @@ export default function MapTooltip() {
     y
   ] = [
     rightSide ? 'right' : 'left',
-    rightSide ? window.innerWidth - currentHoverTarget.x : currentHoverTarget.x,
+    rightSide ? window.innerWidth - currentHoverTarget.x - (widgetsOnRight*400) : currentHoverTarget.x,
     bottomSide ? 'bottom' : 'top',
-    bottomSide ? window.innerHeight - currentHoverTarget.y - 50: currentHoverTarget.y + 50,
+    bottomSide ? window.innerHeight - currentHoverTarget.y - 50: currentHoverTarget.y,
   ]
 
   return (
